Consolidate viewer loading APIs and hull controls
Move file loading into Structure and Trajectory and use source consistently across viewers. Preserve ownership and disposal of loaded versus borrowed data while removing the separate file-viewer components.

Keep hull data, selection, and controls in ConvexHull, with private renderers responsible for geometry. Remove redundant exports and migrate widgets, demos, documentation, and regression tests to the consolidated APIs.
