+++ src/lib/plot/bar/index.ts
@@ -6,7 +6,6 @@ import type StructureBarPlot from './StructureBarPlot.svelte'
 export { default as BarPlot } from './BarPlot.svelte'
 export { default as BarPlotControls } from './BarPlotControls.svelte'
 export { default as SpacegroupBarPlot } from './SpacegroupBarPlot.svelte'
-export { default as StructureBarPlot } from './StructureBarPlot.svelte'
 
 // Props a StructureBarPlot wrapper (BondAnglePlot, CoordinationBarPlot) takes: everything the
 // shell accepts minus what the wrapper derives from `structures` itself.
