+++ src/lib/structure/StructureGallery.svelte
@@ -610,6 +610,7 @@
             {#if idx >= mount_start && idx < mount_end}
               <!-- fill-the-card overrides for Structure's standalone defaults -->
               <Structure
+                allow_file_drop={false}
                 structure={item.structure}
                 {show_controls}
                 scene_props={structure_scene_props}
