+++ src/routes/(demos)/convex-hull/+page.svelte
@@ -10,8 +10,7 @@
     PhaseData,
   } from '$lib/convex-hull'
   import {
-    ConvexHull2D,
-    ConvexHullCanvas,
+    ConvexHull,
     ConvexHullStats,
     GAS_SPECIES,
     process_hull_for_stats,
@@ -35,12 +34,12 @@
   const loaded_data = new SvelteMap<string, PhaseData[]>()
 
   // State for the 3D example with stats display
-  let stats_hull = $state<ReturnType<typeof ConvexHullCanvas>>()
+  let stats_hull = $state<ReturnType<typeof ConvexHull>>()
   const model = $derived(stats_hull?.get_model())
   let max_hull_dist_show_phases = $state(0.5)
 
   // State for the side-by-side stats demo
-  let side_hull = $state<ReturnType<typeof ConvexHullCanvas>>()
+  let side_hull = $state<ReturnType<typeof ConvexHull>>()
   const side_model = $derived(side_hull?.get_model())
   let clicked_entry_id = $state<string | undefined>(undefined)
   let selected_quinary_path = $state<string>(``)
@@ -403,7 +402,7 @@
     <h2>Ternary Chemical Systems</h2>
     <div class="ternary-grid">
       {#each ternary_examples as { title, entries } (title)}
-        <ConvexHullCanvas dim={3} {entries} controls={{ title }} />
+        <ConvexHull {entries} controls={{ title }} />
       {/each}
     </div>
   </section>
@@ -414,8 +413,7 @@
     <div class="quaternary-grid">
       {#each [...loaded_data.entries()].filter( ([p]) => p.includes(`quaternaries`) ) as [path, data] (path)}
         {@const title = hull_system_name(path)}
-        <ConvexHullCanvas
-          dim={4}
+        <ConvexHull
           entries={entries_map.get(path) || data}
           controls={{ title }}
           on_file_drop={handle_file_drop(path)}
@@ -428,7 +426,7 @@
     <h2>Binary Chemical Systems</h2>
     <div class="binary-grid">
       {#each binary_examples as { title, entries } (title)}
-        <ConvexHull2D {entries} controls={{ title }} style="height: 500px" />
+        <ConvexHull {entries} controls={{ title }} style="height: 500px" />
       {/each}
     </div>
   </section>
@@ -443,8 +441,7 @@
       </p>
       {@render feature_list(stats_features)}
       <div class="stats-example-grid">
-        <ConvexHullCanvas
-          dim={3}
+        <ConvexHull
           entries={na_fe_o_entries}
           controls={{ title: `Na-Fe-O with Stats` }}
           bind:this={stats_hull}
@@ -465,8 +462,7 @@
       <h3>Side-by-Side Layout</h3>
       {@render feature_list(side_by_side_features)}
       <div class="side-by-side-example">
-        <ConvexHullCanvas
-          dim={3}
+        <ConvexHull
           entries={li_co_ni_o_data}
           controls={{ title: `Li-Co-O` }}
           bind:this={side_hull}
@@ -488,7 +484,7 @@
       <h2>Highlighted Entries</h2>
       {@render feature_list(highlighted_features)}
       <div class="highlight-grid">
-        <ConvexHull2D
+        <ConvexHull
           entries={binary_examples[1]?.entries ?? []}
           controls={{ title: `Fe-O (${highlighted_fe_o.length} highlighted)` }}
           highlighted_entries={highlighted_fe_o}
@@ -499,8 +495,7 @@
             pulse_speed: 4,
           }}
         />
-        <ConvexHullCanvas
-          dim={3}
+        <ConvexHull
           entries={na_fe_o_entries}
           controls={{ title: `Na-Fe-O (${highlighted_na_fe_o.length} highlighted)` }}
           highlighted_entries={highlighted_na_fe_o}
@@ -511,8 +506,7 @@
             pulse_speed: 3,
           }}
         />
-        <ConvexHullCanvas
-          dim={4}
+        <ConvexHull
           entries={li_co_ni_o_quaternary}
           controls={{ title: `Li-Co-Ni-O (${highlighted_li_co_ni_o.length} highlighted)` }}
           highlighted_entries={highlighted_li_co_ni_o}
@@ -535,8 +529,7 @@
           <div class="marker-legend">
             {#each magnetic_marker_legend as item (item)}<span>{item}</span>{/each}
           </div>
-          <ConvexHullCanvas
-            dim={3}
+          <ConvexHull
             entries={magnetic_ternary_entries}
             controls={{ title: `Na-Fe-O Magnetic Orderings` }}
           />
@@ -545,7 +538,7 @@
           <div class="marker-legend">
             {#each crystallinity_marker_legend as item (item)}<span>{item}</span>{/each}
           </div>
-          <ConvexHull2D
+          <ConvexHull
             entries={crystallinity_entries}
             entry_category={crystallinity_category}
             controls={{ title: `Co-O Crystallinity (custom entry_category)` }}
@@ -561,18 +554,13 @@
       <h2>Temperature-Dependent Free Energies</h2>
       {@render feature_list(temp_features)}
       <div class="temp-grid">
-        <ConvexHull2D
+        <ConvexHull
           entries={temp_binary_entries}
           controls={{ title: `Li-Fe with G(T)` }}
           style="height: 500px"
         />
-        <ConvexHullCanvas
-          dim={3}
-          entries={temp_ternary_entries}
-          controls={{ title: `Li-Fe-O with G(T)` }}
-        />
-        <ConvexHullCanvas
-          dim={4}
+        <ConvexHull entries={temp_ternary_entries} controls={{ title: `Li-Fe-O with G(T)` }} />
+        <ConvexHull
           entries={temp_quaternary_entries}
           controls={{ title: `Li-Fe-Ni-O with G(T)` }}
         />
@@ -593,15 +581,14 @@
         </select>
       </div>
       <div class="gas-grid">
-        <ConvexHull2D
+        <ConvexHull
           entries={gas_demo_fe_o_entries}
           controls={{ title: `Fe-O with ${selected_demo_gas} Pressure` }}
           gas_config={gas_demo_config}
           bind:gas_pressures={gas_demo_pressures}
           style="height: 500px"
         />
-        <ConvexHullCanvas
-          dim={3}
+        <ConvexHull
           entries={gas_demo_ternary_entries}
           controls={{ title: `Fe-Ni-O with ${selected_demo_gas} Pressure` }}
           gas_config={gas_demo_config}
