+++ tests/playwright/convex-hull/convex-hull-2d.test.ts
@@ -99,7 +99,9 @@ test.describe(`ConvexHull2D (Binary)`, () => {
     const before = await get_visible_unstable()
 
     // Set threshold to 0
-    const number_input = controls.getByLabel(`Points threshold (eV/atom)`)
+    const number_input = controls.getByRole(`spinbutton`, {
+      name: `Points threshold (eV/atom)`,
+    })
     // pd2d IS the scatter element (has class "scatter convex-hull-2d")
     const markers = pd2d.locator(`path.marker`)
     const count_before = await markers.count()
